habiba_fixMovieDetail Loading

+++ app/src/main/java/com/example/movieapplication/model/Movie.kt
@@ -4,7 +4,9 @@ data class Movie(
     val id: Int,
     val title: String,
     val overview: String,
-    val posterRes: Double = 0.0,
+    //val posterRes: Double = 0.0,
+    val posterRes: Int = 0,
+
     val poster_path: String,
     val voteAverage: Double = 0.0,
     val releaseDate: String = ""
